Static cast sizes per review

+++ Drv/Ip/TcpServerSocket.cpp
@@ -122,14 +122,16 @@ SocketIpStatus TcpServerSocket::openProtocol(SocketDescriptor& socketDescriptor)
 FwSignedSizeType TcpServerSocket::sendProtocol(const SocketDescriptor& socketDescriptor,
                                                const U8* const data,
                                                const FwSizeType size) {
-    return static_cast<FwSignedSizeType>(::send(socketDescriptor.fd, data, size, SOCKET_IP_SEND_FLAGS));
+    return static_cast<FwSignedSizeType>(
+        ::send(socketDescriptor.fd, data, static_cast<size_t>(size), SOCKET_IP_SEND_FLAGS));
 }
 
 FwSignedSizeType TcpServerSocket::recvProtocol(const SocketDescriptor& socketDescriptor,
                                                U8* const data,
                                                const FwSizeType size) {
     // recv will return 0 if the client has done an orderly shutdown
-    return static_cast<FwSignedSizeType>(::recv(socketDescriptor.fd, data, size, SOCKET_IP_RECV_FLAGS));
+    return static_cast<FwSignedSizeType>(
+        ::recv(socketDescriptor.fd, data, static_cast<size_t>(size), SOCKET_IP_RECV_FLAGS));
 }
 
 }  // namespace Drv
